docs(ComboBox): update creevey test
