+++ packages/react-ui/.creevey/images/ComboBox/with featureFlags value change/Update value/firefox/updatedValue.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:479421bece79b5a198bc870507be7bd94e49f5a35e4f59040f3bc6eebc54da74
+size 19341
