+++ packages/react-ui/.creevey/images/ComboBox/with featureFlags value change/Update value/firefox/withTypedText.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:668293968e3d7101457a9dc87a6f3acf0d8afda4ca1687e48a1fd57af179d822
+size 21843
