+++ packages/react-ui/.creevey/images/ComboBox/with featureFlags value change/Update value/firefox2022/updatedValue.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:0a5bcfe2e4b122316de12793db9df204c447b15206bbac3bc5c5ee7d3ddb5100
+size 18525
