+++ packages/react-ui/.creevey/images/ComboBox/with featureFlags value change/Update value/firefox2022/withTypedText.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:ec8605a4f7e7b0afc4944b4bcf6392904a173567c598368d3b371efd2028ecbe
+size 23773
