+++ packages/react-ui/.creevey/images/ComboBox/with featureFlags value change/Update value/firefox2022Dark/updatedValue.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:f79b97b431beedc935e8a53469b20390cdd9adeac8e3f1904330d7c002e1673f
+size 18531
