+++ packages/react-ui/.creevey/images/ComboBox/with featureFlags value change/Update value/firefox2022Dark/withTypedText.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:7741d8121ac29f4ba11735da355215315bc58798ec8a61360af123494f1918a2
+size 21781
