+++ packages/react-ui/.creevey/images/ComboBox/with featureFlags value change/Update value/firefox8px/updatedValue.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:98c27c945b3bc1538f125009fa8e1ca8ac8447f16da7e766b59bfbaba3f09089
+size 19554
