+++ packages/react-ui/.creevey/images/ComboBox/with featureFlags value change/Update value/firefox8px/withTypedText.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:fa58386d4849104c617300b9b41c89cd4570b63a83b5f5cf6b64e0f222091912
+size 22160
