+++ packages/react-ui/.creevey/images/ComboBox/with featureFlags value change/Update value/firefoxDark/updatedValue.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:5034b1e8ac7cb20b010c61449a292ee6cd6d03cf277f1fa50dd31f878be3ea22
+size 19289
