+++ packages/react-ui/.creevey/images/ComboBox/with featureFlags value change/Update value/firefoxDark/withTypedText.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:da9c1d26044641f59bcb0c05a09e9c59f519359a1db69de9e3059f8b32a6c9d6
+size 20997
